uint_8

+++ include/server/Arbitre.hpp
@@ -24,7 +24,7 @@ public:
 
 	void	put_stone_game(ITCP_protocol<Client> &itcp_protocol, ITCP_protocol<Client>::Game_stone *stone);
 
-	bool	can_capture(ITCP_protocol<Client>::Game_stone * stone, int coord[8][4]) const;
+	bool	can_capture(ITCP_protocol<Client>::Game_stone * stone, uint8_t coord[8][4]) const;
 
 	bool	can_put_stone(ITCP_protocol<Client>::Game_stone *stone) const;
 
@@ -46,11 +46,11 @@ public:
 
 	bool is_diag_no_iso_three(int coord[4], ITCP_protocol<Client>::Game_stone * stone) const;
 
-	const ITCP_protocol<Client>::Game_stone::Color& operator()(int x, int y) const;
+	const ITCP_protocol<Client>::Game_stone::Color& operator()(unsigned int x, unsigned int y) const;
 
-	ITCP_protocol<Client>::Game_stone::Color& operator()(int x, int y);
+	ITCP_protocol<Client>::Game_stone::Color& operator()(unsigned int x, unsigned int y);
 
-	static const int board_size = 19;
+	static const uint8_t board_size = 19;
 private:
 	std::array<ITCP_protocol<Client>::Game_stone::Color, board_size * board_size>	m_board;
 	int			m_white_loose;
